Refactor MapsActivity
